storage-client: absorb ComputeInstanceId
Part of the cluster unification epic (MaterializeInc/cloud#4929).

Move the ComputeInstanceId type to the storage-client crate and rename
it to StorageClusterId. Then re-export it as ComputeInstanceId from
compute-client.

This will facilitate creating storage clusters independently of
ingestions and exports.
